+++ rust/kadena/kadena-client/src/models/misc/continue_transfer_remote_dto.rs
@@ -7,9 +7,9 @@ pub struct ContinueTransferRemoteDto {
     host: String,
     network: String,
     #[serde(rename = "chain_id")]
-    chain_id: u8,
+    chain_id: u16,
     pact_id: String,
-    destination_chain_id: u8,
+    destination_chain_id: u16,
     step: u8,
     rollback: bool,
 }
@@ -18,14 +18,14 @@ impl ContinueTransferRemoteDto {
     pub fn new(
         conf: &ChainwebConf,
         pact_id: String,
-        destination_chain_id: u8,
+        destination_chain_id: u16,
         step: u8,
         rollback: bool,
     ) -> Self {
         Self {
             host: conf.url.to_string(),
             network: conf.network_id.to_string(),
-            chain_id: conf.chain_id as u8,
+            chain_id: conf.chain_id as u16,
             pact_id,
             destination_chain_id,
             step,
